Add TLS client and server wildcard certificates

+++ base/source/ESBString.cpp
@@ -4,14 +4,14 @@
 
 namespace ESB {
 
-ESB::UInt32 StringHash(const char *str) {
+ESB::UInt64 StringHash(const char *str) {
   assert(str);
   if (!str) {
     return 0U;
   }
 
   // From Sedgewick's Algorithms in C, 4th ed, p.579
-  ESB::UInt32 h = 0U, a = 31415, b = 27183, M = 7919;
+  ESB::UInt64 h = 0U, a = 31415, b = 27183, M = 7919;
   for (const char *v = str; *v; v++, a = a * b % (M - 1)) {
     h = (a * h + *v) % M;
   }
